refactor main queue
