+++ aux_ui/lib/screens/main/main_queue.dart
@@ -1,20 +1,15 @@
-import 'dart:typed_data';
-
 import 'package:aux_ui/aux_lib/song.dart';
 import 'package:aux_ui/aux_lib/spotify_session.dart';
-import 'package:aux_ui/routing/router.dart';
-import 'package:aux_ui/widgets/buttons/queue_item_action.dart';
-import 'package:aux_ui/widgets/layout/aux_card.dart';
-import 'package:aux_ui/widgets/layout/queue_item.dart';
-import 'package:aux_ui/widgets/layout/song_countdown.dart';
+import 'package:aux_ui/widgets/queue_main_display/current_song.dart';
+import 'package:aux_ui/widgets/queue_main_display/queue_header.dart';
+import 'package:aux_ui/widgets/queue_main_display/song_countdown.dart';
 import 'package:aux_ui/widgets/layout/song_list.dart';
+import 'package:aux_ui/widgets/queue_main_display/song_up_next.dart';
 import 'package:flutter/material.dart';
 import 'package:aux_ui/theme/aux_theme.dart';
 import 'package:aux_ui/widgets/layout/queue_container.dart';
-import 'package:aux_ui/widgets/layout/playback_controls.dart';
-import 'package:spotify_sdk/models/image_uri.dart';
+import 'package:aux_ui/widgets/queue_main_display/playback_controls.dart';
 import 'package:spotify_sdk/models/player_state.dart';
-import 'package:spotify_sdk/models/track.dart';
 
 class MainQueue extends StatefulWidget {
   final SpotifySession spotifySession;
@@ -24,14 +19,29 @@ class MainQueue extends StatefulWidget {
   _MainQueueState createState() => _MainQueueState();
 }
 
+class _ExpandQueueButton extends StatelessWidget {
+  const _ExpandQueueButton({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ButtonTheme(
+        height: SizeConfig.blockSizeVertical * 3,
+        child: OutlineButton(
+            padding: EdgeInsets.only(top: 5, bottom: 5, right: 7, left: 7),
+            borderSide: BorderSide(color: auxAccent),
+            onPressed: () {},
+            color: Colors.transparent,
+            child: Row(children: <Widget>[
+              Icon(Icons.unfold_more,
+                  color: auxAccent, size: 10.0, semanticLabel: "expand queue"),
+              Text("view full queue", style: auxCaption)
+            ])));
+  }
+}
+
 class _MainQueueState extends State<MainQueue> {
   List<Song> yourSongs;
   List<Song> queueSongs;
-  Widget _currPlaying;
-  Widget _expandQueue;
-  Widget _songUpNext;
-  Widget _header;
-  bool _initialized = false;
 
   @override
   void initState() {
@@ -55,148 +65,8 @@ class _MainQueueState extends State<MainQueue> {
     });
   }
 
-  void _initializeWidgets() {
-    if (_initialized) return;
-    _initialized = true;
-    _setSongUpNext();
-    _setExpandQueue();
-    _setHeader();
-  }
-
-  Widget _getCurrPlaying(PlayerState playerState) {
-    // TODO possibly pull out into another component
-
-    Track track = playerState.track;
-    String name = track.name;
-    String artist = track.artist.name;
-    ImageUri imageUri = track.imageUri;
-    String contributor = "Diane"; // TODO: don't hardcode
-    double progress = playerState.playbackPosition / track.duration;
-
-    Widget right = QueueItemAction(onPressed: () {}, icons: [
-      Icon(
-        Icons.favorite_border,
-        color: auxAccent,
-        size: 16.0, // TODO: scale
-        semanticLabel: "original song",
-      ),
-      Icon(
-        Icons.favorite,
-        color: auxAccent,
-        size: 16.0, // TODO: scale
-        semanticLabel: "liked song",
-      )
-    ]);
-
-    return FutureBuilder(
-      future: spotifySession.getImage(imageUri),
-      builder: (BuildContext context, AsyncSnapshot<Uint8List> snapshot) {
-        if (snapshot.hasData) {
-          var albumCover = snapshot.data;
-          return AuxCard(
-              borderColor: auxBlurple,
-              padding: 15.0,
-              child: Column(
-                children: <Widget>[
-                  QueueItem(
-                    song: new Song(name, artist, albumCover, contributor),
-                    showContributor: true,
-                    rightPress: right,
-                    isAccent: true,
-                  ),
-                  Padding(
-                      padding: EdgeInsets.only(top: 10),
-                      child: LinearProgressIndicator(
-                          value: progress,
-                          backgroundColor: auxDDGrey,
-                          valueColor: new AlwaysStoppedAnimation<Color>(auxBlurple)))
-                ],
-              ));
-        } else {
-          return Center(
-            child: Text("Getting image"), // TODO: replace
-          );
-        }
-      }
-    );
-  }
-
-  void _setExpandQueue() {
-    _expandQueue = ButtonTheme(
-        height: SizeConfig.blockSizeVertical * 3,
-        child: OutlineButton(
-            padding: EdgeInsets.only(top: 5, bottom: 5, right: 7, left: 7),
-            borderSide: BorderSide(color: auxAccent),
-            onPressed: () {},
-            color: Colors.transparent,
-            child: Row(children: <Widget>[
-              Icon(Icons.unfold_more,
-                  color: auxAccent, size: 10.0, semanticLabel: "expand queue"),
-              Text("view full queue", style: auxCaption)
-            ])));
-  }
-
-  void _setSongUpNext() {
-    Widget right = QueueItemAction(onPressed: () {}, icons: [
-      Icon(Icons.more_vert,
-          color: auxAccent,
-          size: 20.0, // TODO: scale
-          semanticLabel: "get next song")
-    ]);
-
-    _songUpNext = QueueItem(
-      song: queueSongs[0],
-      showContributor: true,
-      rightPress: right,
-    );
-  }
-
-  int getNumInParty() {
-    // TODO: implement
-    return 12;
-  }
-
-  String getHost() {
-    // TODO: implement
-    return "Diane";
-  }
-
-  Widget _getHeaderChip(IconData iconName, String text, Color color) {
-    return Container(
-        padding: EdgeInsets.only(right: 16), // TODO: scale or finalize
-        child: Row(
-          children: <Widget>[
-            Icon(iconName, color: color, size: 10),
-            Padding(
-                padding: EdgeInsets.only(left: 1),
-                child: Text(text, style: auxBody1))
-          ],
-        ));
-  }
-
-  void _setHeader() {
-    _header = Column(children: <Widget>[
-      Align(
-          alignment: Alignment.bottomLeft,
-          child: Text('too queue for u', style: auxDisp2)),
-      Padding(
-          padding: EdgeInsets.only(top: 3), // TODO: scale or finalize
-          child: Row(
-            children: <Widget>[
-              _getHeaderChip(Icons.fiber_manual_record, "LIVE", Colors.red),
-              _getHeaderChip(
-                  Icons.group, "${getNumInParty()} people in group", auxAccent),
-              _getHeaderChip(
-                  Icons.person_outline, "hosted by ${getHost()}", auxAccent)
-            ],
-          ))
-    ]);
-  }
-
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context);
-    _initializeWidgets();
     return StreamBuilder<PlayerState>(
         stream: widget.spotifySession.getPlayerState(),
         initialData: PlayerState(null, true, 1, 1, null, null),
@@ -217,12 +87,12 @@ class _MainQueueState extends State<MainQueue> {
                               Container(
                                   padding: EdgeInsets.only(
                                       left: 12, right: 12, top: 42, bottom: 8),
-                                  child: _header),
-                              _getCurrPlaying(playerState),
+                                  child: QueueHeader()),
+                              CurrentSong(playerState: playerState, spotifySession: widget.spotifySession),
                               QueueContainer(
                                   title: 'up next',
-                                  child: _songUpNext,
-                                  titleWidget: _expandQueue),
+                                  child: SongUpNext(song: queueSongs[0]),
+                                  titleWidget: const _ExpandQueueButton()),
                               QueueContainer(
                                 title: 'your songs',
                                 child:
