+++ aux_ui/lib/screens/main/main_search.dart
@@ -1,7 +1,7 @@
 import 'package:aux_ui/aux_lib/song.dart';
 import 'package:aux_ui/widgets/buttons/queue_item_action.dart';
 import 'package:aux_ui/widgets/layout/queue_item.dart';
-import 'package:aux_ui/widgets/layout/song_countdown.dart';
+import 'package:aux_ui/widgets/queue_main_display/song_countdown.dart';
 import 'package:aux_ui/widgets/layout/song_list.dart';
 import 'package:flutter/material.dart';
 import 'package:aux_ui/theme/aux_theme.dart';
